update LICENCE, some warning fixes

+++ src/Preconditioners.h
@@ -1480,7 +1480,7 @@ public:
 #ifdef HAVE_OPENMP
 #pragma omp parallel for
 #endif
-    for (size_t i = 0; i < x.size(); ++i) {
+    for (int i = 0; i < x.size(); ++i) {
       x[i] = b[i];
     }
